# app/Http/Resources/CourseResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Contracts\Support\Arrayable;
use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

class CourseResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @param  Request  $request
     * @return array
     */
    public function toArray($request): array
    {
        return [
            'id' => $this->id,
            'name' => $this->name,
            'image' => env('APP_STORAGE') . $this->image,
            'background' => env('APP_STORAGE') . $this->background,
            'video_counts' => $this->videos->count(),
        ];
    }
}