# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mohir kids Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "bakhriddin.inamov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://mohir-kids.test/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/course/list": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get list of courses",
                "description": "Get a list of courses",
                "operationId": "courseList",
                "parameters": [
                    {
                        "name": "bearer",
                        "in": "header",
                        "description": "Children access token",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/course/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get course",
                "description": "Get a single course",
                "operationId": "getCourse",
                "parameters": [
                    {
                        "name": "bearer",
                        "in": "header",
                        "description": "Children access token",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CourseResource": {
                "title": "CourseResource",
                "description": "Course resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CourseResource"
                }
            },
...
        }
    }
}